requirements.txt
discord.py>=2.3
python-dotenv>=1.0
requests>=2.32
shapely>=2.0
matplotlib>=3.8
